requirements.txt
requests
discord.py
Pillow
python-Levenshtein